refactor: token 서비스로 로직 단순화
- #10
